requirements.txt
asyncio==3.4.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pendulum==2.1.0
pygal==2.4.0
python-dateutil==2.8.1
pytzdata==2019.3
requests==2.23.0
six==1.14.0
smmap==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
